Player body now completely protected by back shield
I've done it!  Signal suppression functionality assigns a higher
priority to energy absorptions from the shield compared to the body, so
in the case of simultaneous signalling, the body signal is ignored and
only the shield signal is let through: the player does not take damage.  Also made some changes - added player shield and enemy unblockable physics layers, added script that disables all energy sources that touch the player shield