Working towards hold to charge, tap to shoot mechanic
Improved stock change over time script to allow requests of other scripts to start and stop the stock change, at the expense of the auto-regeneration of the shield durability, which will have to be accomplished with a trivial script that starts the stock change over time as soon as the scene starts.  This allows holding down the mouse button to gradually increase the gun charge rate.  Currently, holding down the mouse button fires the gun - need to change to when the mouse button is released.  Also... theres a bug.  Once the PowerLevelEmitter tries to fire the charge gun, the program freezes.  I already checked the emissionRate on the charge gun, so its not like its trying to emit multiple times in a frame.  Try peppering the area around the emit call in PowerLevelEmitter with debug breaks or something