Enormous refactor to event systems for scripts that publish events
All scripts that publish important events now use the Event class which has a programmatic event that passes in an argument that gives data about the event and a UnityEvent which takes no arguments and gives no info about the event but can be set conveniently in the editor without adding a single line of code.  A large volumn of scripts have been marked as deprecated as a result of the new event system, most of them having the keyword On.  These scripts had no interest in the actual data of the event and simply called a completely unrelated function, thus they are easily replaced by adding the function of interest to the UnityEvent in the editor.  Initial tests good, Im apprehensive about the player shields, though, since that involved lots of the scripts that are now deprecated.  Also, the player prefab is the only one that has been update.  None of the enmies have been updated with the new event systems, and all should have at least one runtime error of a missing State script.  Speaking of which, State scripts should now be added to game objects manually rather than programmatically constructed so that functions can be added to the UnityEvent in the UI