ManualTierSelector editor is much easier to work with
The ManualTierSelector is not getting the current level from a static variable, so it keeps resetting the level back to zero even though the level increments when the player is victorious.  The next step is to design the level so that it persists through scene loading.  Also, an IndexOutOfRange exception is thrown whenever the number of tier selections is set to 7 or higher, so that has to be fixed as well if we want to have more than 7 stages (and we definitely do)