Improved LabelledComponentKeeper to be a MonoBehaviour with intuitive editor
Currently, the spawners are not triggered properly when the current level is advanced by one.  Part of the reason for this may be that the CyurrentLevel prefab creates a DontDestroyOnLoad copy of itself ever single time that the scene reloads.  First we need to make sure only one CurrentLevel object is set up to persist through scene loading, and that each subsequent attempt to create a persistent copy of the object is ignored.  After that, we can test to see if that is the issue with the spawner.  The next thing to do after that would be to change the tier selector such that the duration is taken into account, whereas it is currently being ignored