Tier selection duration works great
Currently, no level selections occur at level 6 or beyond for no obvious reason. I hypothesize that this is no coincidence with the fact that the editor throws an ArgumentOutOfRange exception whenever the ManualTierSelectorEditor tries to go to level 6 or above.  Next, the editor issue with the ArgumentOutOfRange problem should be fixed